fix: implement auto linking on register
